+++ app/src/atoms/SoftwareKeyboard/NumericalKeyboard/NumericalKeyboard.stories.tsx
@@ -2,7 +2,6 @@ import { useRef, useState } from 'react'
 
 import {
   DIRECTION_COLUMN,
-  Flex,
   InputField,
   POSITION_ABSOLUTE,
   SPACING,
@@ -43,7 +42,12 @@ const Keyboard = (args): JSX.Element => {
   const [value, setValue] = useState<string>('')
   const keyboardRef = useRef(null)
   return (
-    <Flex flexDirection={DIRECTION_COLUMN} gridGap={SPACING.spacing16}>
+    <div
+      style={{
+        flexDirection: DIRECTION_COLUMN,
+        gridGap: SPACING.spacing16,
+      }}
+    >
       <form id="test_form">
         <InputField
           value={value}
@@ -54,11 +58,13 @@ const Keyboard = (args): JSX.Element => {
           }}
         />
       </form>
-      <Flex
-        position={POSITION_ABSOLUTE}
-        top="20%"
-        width="22.5rem"
-        height="max-content"
+      <div
+        style={{
+          position: POSITION_ABSOLUTE,
+          top: '20%',
+          width: '22.5rem',
+          height: 'max-content',
+        }}
       >
         {showKeyboard && (
           <NumericalKeyboard
@@ -69,8 +75,8 @@ const Keyboard = (args): JSX.Element => {
             hasHyphen={hasHyphen}
           />
         )}
-      </Flex>
-    </Flex>
+      </div>
+    </div>
   )
 }
