+++ app/src/atoms/SoftwareKeyboard/index.ts
@@ -1,10 +1,8 @@
 export { AlphanumericKeyboard } from './AlphanumericKeyboard'
 export { IndividualKey } from './IndividualKey'
 export { FullKeyboard } from './FullKeyboard'
-export {
-  NumericalKeyboard,
-  StatelessNumericalKeyboard,
-  applyNumericalKeyboardKey,
-  isValidNumericalInput,
-} from './NumericalKeyboard'
-export type { NumericalKeyboardKey } from './NumericalKeyboard'
+export { NumericalKeyboard } from './NumericalKeyboard'
+export { StatelessNumericalKeyboard } from './NumericalKeyboard/StatelessNumericalKeyboard'
+export { applyNumericalKeyboardKey } from './utils/applyNumericalKeyboardKey'
+export { isValidNumericalInput } from './utils/isValidNumericalInput'
+export type { NumericalKeyboardKey } from './types'
